feat: integrated gossipsub metrics with cli runtime
