+++ examples/metrics/runner.py
@@ -3,6 +3,9 @@ import trio
 
 from examples.metrics.coordinator import COMMANDS, Node
 from libp2p.metrics.metrics import Metrics
+from libp2p.tools.async_service.trio_service import (
+    background_trio_service,
+)
 from libp2p.utils.address_validation import get_available_interfaces
 
 
@@ -15,34 +18,39 @@ async def main() -> None:
         node.host.run(listen_addrs=listen_addrs),
         trio.open_nursery() as nursery,
     ):
-        # Start metrics service
-        metrics = Metrics()
-
-        nursery.start_soon(
-            metrics.start_prometheus_server, node.host.metric_recv_channel
-        )
         nursery.start_soon(node.host.get_peerstore().start_cleanup_task, 60)
-        nursery.start_soon(node.command_executor, nursery)
-
         print(f"Host multiaddr: {node.host.get_addrs()[0]}")
-        await trio.sleep(1)
 
-        print("Entering intractive mode, type commands below.")
-        promt_session = PromptSession()
-        print(COMMANDS)
-
-        while not node.termination_event.is_set():
-            try:
-                _ = await trio.to_thread.run_sync(input)
-                user_input = await trio.to_thread.run_sync(
-                    lambda: promt_session.prompt("Command> ")
-                )
-                cmds = user_input.strip().split(" ", 2)
-                await node.input_send_channel.send(cmds)
-
-            except Exception as e:
-                print(f"Error in the interactive shell: {e}")
+        async with background_trio_service(node.pubsub):
+            async with background_trio_service(node.gossipsub):
                 await trio.sleep(1)
+                await node.pubsub.wait_until_ready()
+                print("Gossipsub and Pubsub services started !!")
+
+                # METRICS
+                metrics = Metrics()
+                nursery.start_soon(
+                    metrics.start_prometheus_server, node.host.metric_recv_channel
+                )
+                nursery.start_soon(node.command_executor, nursery)
+                await trio.sleep(1)
+
+                print("Entering intractive mode, type commands below.")
+                promt_session = PromptSession()
+                print(COMMANDS)
+
+                while not node.termination_event.is_set():
+                    try:
+                        _ = await trio.to_thread.run_sync(input)
+                        user_input = await trio.to_thread.run_sync(
+                            lambda: promt_session.prompt("Command> ")
+                        )
+                        cmds = user_input.strip().split(" ", 2)
+                        await node.input_send_channel.send(cmds)
+
+                    except Exception as e:
+                        print(f"Error in the interactive shell: {e}")
+                        await trio.sleep(1)
 
     print("Shutdown complete, Goodbye!")
